Add test env variable
I've added a TEST env variable so that the tests can pass. In tests, we mock certain things, like add_action calls, so certain paths of the commands will be taken (certain branches). In order for us to test if the functionality is working during tests, we need to identify these cases. For now I only found two of such cases, where this is needed, but I don't think these will be the only two.