Wrap mocks in a function that can be initialized on every test
This was done so that we can call the brain monkey mock methods where we need them. Because pest is a unique wrapper of phpunit, it handles things in a different way, so the Helpers.php will be 'cleaned' out by brainmonkey on every run, causing failures. This way we can just reinitialize them in an easy way, and reduce duplications.