Test CLI part (#160)
* Add a way to mock system calls

Had to override shell_exec in order to prevent running shell command during testing

* Add project init test

* Extract command name to a constant

Reduces duplication

* Fix indentation in one test

* Add additional system calls to mock

* Add coverage ignore to reflection exception

* Add new cli tests and fix db exports test

* Replace exit() with WP-CLI error

* Remove unused command

* Add CLI tests

* Ignore errors about undefined reflection classes

* Remove unreachable code and code ignore exit() statement

* Add cli tests and part of the abstract cli tests

* Update the phpcs to test for PHP 7.3

* Add mockery mock helper

* Remove unnecessary polyfill, fix the exit codes

* Add and improve CLI tests

* Replace explicit calls to Mockery with a helper

Makes it neater

* Remove breaking tests and test renameUseFrontendLibs

* Add dummy data for blocks tests

* Test blocks

I've fixed the failing tests because I needed to check the CLI method I've changed - getExampleTemplate. I'll need to test this further with the blocks tests, as they seem to be the only ones that are adding the third parameter as true (which affects the changes in the code).

* Test blocks

I've fixed the failing tests because I needed to check the CLI method I've changed - getExampleTemplate. I'll need to test this further with the blocks tests, as they seem to be the only ones that are adding the third parameter as true (which affects the changes in the code).

* Refactor the getExampleTemplate method

* Typo fix

* Refactor the getProjectConfigRootPath so that the dir is correct when testing

* Update test for CLI and update

* Fix PR comments

* Return the load command for the Cli class

* Add test for load method