Cleanup base test class
Add wpcli mock as a property so that we can alias new runcommand commands in individual tests.

The initial runcommand mock is removed from the base test class, so that we can alias is in each test where we need it to be different, based on the arguments. If we leave it in the base test class the alias is set up before every
test, and we cannot 'redefine' it inside the test class.

The property has to be protected in order for it to be available in every test case.