Update libs with wpcs 3 changes [beta] (#373)
* Update composer dependencies

* Update phpcs config

* Remove unused ignored error

* Fix deprecation notice

Depth parameter shouldn't be set to 0.

* Fix the tests failures because of inability to check the list of items vs source items

* Remove extra space from init blocks command

* Update readme

* Fix phpcs violation

* Fix deprecations with using traits directly

Added anonymous classes instead.

* Add mocks for the failing tests

* Fix test deprecations

* Update the gh action workflow

* Workflow fix

* Workflow update

8.3 allow failure for non-released version of PHP.

* Minor formatting fix on helpers

* Replace assertions with expectations API

Should prevent failures on PHPUnit <9 where assertObjectHasProperty isn't defined.

* Prevent the dynamic properties deprecation notices in tests

This will still be an issue on composer installs with lowest dependencies where dynamic properties isn't allowed/fixed in mockery.

* Minor grammar typo fix in a test

* Update workflow to allow failures on lowest composer dependencies

* Update moveItem method

* Improve type description in moveItem method

* Improve the moveItems method

In real life cases, the sourceItem array will contain list of items in the folder, and the folder as a value, so we need to check that as well. This should probably come up in tests, unless the dataset is not correctly set to mimick how the real life projects are behaving. This needs investigation.