Better CLI loading for ServiceInterface implementations (#438)
* Add ShouldLoadInCliContext class attribute

This commit adds a new class attribute that can be used to indicate a service class (i.e. implementing ServiceInterface) should be loaded in the CLI context

* Add ShouldLoadInCliContext class attribute

This commit adds a new class attribute that can be used to indicate a service class (i.e. implementing ServiceInterface) should be loaded in the CLI context.

* Adapt AbstractMain to ShouldLoadInCliContext

Changes registerServices() so it only additionally loads ShouldLoadInCliContext annotated classes if and only if under WP-CLI.